[ADD] sale_outplacement AFC-1556: adding button for task_ids
