+++ sale_outplacement/models/product_task.py
@@ -20,6 +20,16 @@
 #
 ################################################################################
 
-from . import outplacement
-from . import product_task
-from . import hr_department
+from odoo import api, fields, models
+
+
+class ProjectTask(models.Model):
+    _name = 'product.task'
+
+    outplacement_id = fields.Many2one('outplacement', 'Outplacement')
+    activity_id = fields.Many2one('res.joint_planning', 'Activity')
+    joint_planing_type = fields.Selection([
+        ('kvl', 'KVL'),
+        ('preplaning', 'Preplaning'),
+        ('endraport', 'Endraport'),
+    ], 'JP_Type')
