+++ sale_outplacement/security/ir.model.access.csv
@@ -1,1 +1,2 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_product_task,access_product_task,model_product_task,base.group_user,1,1,1,1
