+++ sale_outplacement/views/hr_department_view.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <odoo>
     <record id="view_department_form" model="ir.ui.view">
-            <field name="name">hr.department.form</field>
-            <field name="model">hr.department</field>
-            <field name="inherit_id" ref="hr.view_department_form"/>
-            <field name="arch" type="xml">
-                <field name="company_id" position="after">
-                    <field name="ka_ref"/>
-                </field>
+        <field name="name">hr.department.form</field>
+        <field name="model">hr.department</field>
+        <field name="inherit_id" ref="hr.view_department_form"/>
+        <field name="arch" type="xml">
+            <field name="company_id" position="after">
+                <field name="ka_ref"/>
             </field>
+        </field>
     </record>
 </odoo>
