+++ sale_outplacement/views/outplacement_view.xml
@@ -5,6 +5,14 @@
             <field name="model">outplacement</field>
             <field name="inherit_id" ref="outplacement.outplacement_form_view"/>
             <field name="arch" type="xml">
+                <div name="button_box" position="inside">
+                    <button name="action_project_task"
+                        class="oe_stat_button"
+                        icon="fa-list-alt"
+                        type="object">
+                        <field name="tasks_count" widget="statinfo" string="Tasks"/>
+                    </button>
+                </div>
                 <group position="after">
                     <group>
                         <group>
